Added Spring-Security configuration. Added JPA support for domain entities

+++ src/main/java/com/udemy/msb/restfulstarter/services/UserService.java
@@ -14,5 +14,5 @@ public interface UserService {
 
     void deleteUserById(Long id);
 
-    int totalUsers();
+    long totalUsers();
 }
